basic types and vectors tree hasher updates
