--- bytes.go
@@ -1,10 +1,0 @@
-package ssz
-
-// ToBytes32 is a convenience method for converting a byte slice to a fix
-// sized 32 byte array. This method will truncate the input if it is larger
-// than 32 bytes.
-func ToBytes32(x []byte) [32]byte {
-	var y [32]byte
-	copy(y[:], x)
-	return y
-}
