+++ examples/example_and_test.go
@@ -1,4 +1,4 @@
-package ssz
+package examples
 
 import (
 	"bytes"
