+++ examples/example_encode_test.go
@@ -1,4 +1,4 @@
-package ssz_test
+package examples_test
 
 import (
 	"bytes"
