+++ ssz_utils_cache.go
@@ -17,7 +17,7 @@ type decoder func(io.Reader, reflect.Value) (uint32, error)
 
 type encodeSizer func(reflect.Value) (uint32, error)
 
-type hasher func(reflect.Value) ([]byte, error)
+type hasher func(reflect.Value) ([32]byte, error)
 
 type sszUtils struct {
 	encoder
